// src/app/page/master/settingdatafixed/settingdatafixed.interface.ts
import { EventEmitter } from '@angular/core';

export interface SettingDataFixed {
	
	alamatEmail?: string;
	faksimile?: string;
	fixedPhone?: string;
	jamBuka?: string;
	jamTutup?: string;
	kdDepartemen?: string;
	kdKelasHead?: string;
	kdModulAplikasi?: string;
	kdPegawaiKepala?: string;
	kode?: string;
	kodeExternal?: string;
	lokasiRuangan?: string;
	mobilePhone?: string;
	namaExternal?: string;
	namaRuangan?: string;
	noCounter?: any;
	noRuangan?: string;
	prefixNoAntrian?: string;
	reportDisplay?: string;
	statusEnabled?: true;

	version?: any;
	noRec?: string;
	label?: any;
	value?: any;
	typeField?: any;
	tabelRelasi?: any;
	fieldKeyTabelRelasi?: any;
	fieldReportDisplayTabelRelasi?: any;
	prefix?: any;
	namaField?: any;
	nilaiFieldDrop?: any;

}